Handle Page Fault in `trap()`
Currently incomplete - needs to get the page fault address. This will likely be found in one of the registers of the trapframe tf. The xv86-64 manual, the textbook, or the syscall man page might have more information.
Added `allocated` field in mmap_entry, which is 1 if memory is allocated and 0 if not. Unmap needs to be revised to deallocate if allocated is 1.