Replace Unused `mmap_entry->allocated` with `mmap_entry->mapping_process`
Used for shared pages only. When a process calls exit(), it will check each entry if mapping_process is true. If so, the memory is freed. This is to prevent the child process from freeing physical pages that are still being referenced by parent processes.

However, tests 25, 26, and 28 still do not pass despite this addition. More revisions and additions to the code will be necessary.

TODO and debug comments are also added. They can be referenced for debugging or removed once the project is finished.

README.md has been updated.