Add Mutex Lock Structure, Functions
Added a `typedef` structure `mutex` in mutex.h, which is identical to the provided `sleeplock` except it does not contain a name.
Added user-level `minit(mutex*)` function in mutex.c (defined in user.h), and system calls `macquire(mutex*)` and `mrelease(mutex*)` in sysproc.c. The implementation of these methods are more or less identical to the `sleeplock` counterparts.
The program compiles and, when I ran the tests twice, passed tests 1-3. Running the tests for more iterations may be helpful to ensure no concurrency bugs are present.
The Makefile has been modified to account for the new mutex.c file.