Fixed Round Robin Policy in Scheduler
Originally, I had forgotten that during a context switch, the scheduler comes back from where it had left off. Thus, the round robin policy was not properly implemented when I had implemented nice values and priority inheritance.
Now, this should be fixed. I added an auxillary function findminnice() to replace the loop finding the minimum elevated nice value in the scheduler, and called it both at the beginning of the scheduler and after every context switch.