Add Python Workload File
This file was added to the starter code to help us generate workloads and test our implementation. (See https://piazza.com/class/lrl9gion4s33ro/post/869 for more info.)
A README.md file was also included to give info for using this file. I copied its contents and pasted it to the end of this README.md file.