Add Partially Complete KV Store Functions
`hashtable` will be the hashtable used to store key-value pairs. Certain components are incomplete and will throw exceptions if attempted to compile. We will need to finish the methods in `kv_store.c`, including the server `main()` method, for it to be functional.
`ring_buffer.c` file is also added, but the functions are not implemented. Change imports as needed.