Create `init_kv_store()` Function, `hashtable` in Server Process
Server `main()` function in `kv_store.c` now initializes the hashtable by calling `init_kv_store()`. It also frees the allocated memory at the end.
Replaced `mmap` with `malloc` for allocating memory for value nodes.
Chaining is used for the hashtable implemented, so resizing would not be necessary.
We still need to create multiple threads for the server process.