Create Thread Function, Threads in Server Main
The server thread function gets requests from the ring and stores the output in the desired location.
The `kv_store` structure no longer contains an indeces array as it is unnecessary, and `value_node` is renamed to `keyvalue_node`.
TODO: Use stack-allocated memory instead of heap-allocated memory for the `kv_store`, and revise obtaining the shared mmap memory pointer. (I am still figuring out how to do the latter.)