Change Bitmap Sizes to Use `i`, `d` Bits Instead of Bytes
Also fixed `root_inode->mode` bitwise operations, and utilized bitwise operations in other places. (The latter is not necessary.)
mkfs.c now correctly checks to make sure the data image file is large enough to contain the file system.
More debug and error print statements are added. The `superblock` in wfs.c points to the superblock in the disk image file, as desired.
Ensured that at least one inode and data block will exist in the file system for the root directory.